fix checkpoint loading after killing backend
