+++ classes.py
@@ -12,7 +12,7 @@ from .property_groups.seamless_result import SeamlessResult
 from .ui.panels import dream_texture, history, upscaling, render_properties
 from .preferences import OpenURL, StableDiffusionPreferences,\
     ImportWeights, Model, ModelSearch, InstallModel, PREFERENCES_UL_ModelList,\
-    Checkpoint, LinkCheckpoint, UnlinkCheckpoint, PREFERENCES_UL_CheckpointList
+    CheckpointGroup, LinkCheckpoint, UnlinkCheckpoint, PREFERENCES_UL_CheckpointList
 
 from .ui.presets import DREAM_PT_AdvancedPresets, DREAM_MT_AdvancedPresets, AddAdvancedPreset, RestoreDefaultPresets
 
@@ -70,7 +70,7 @@ PREFERENCE_CLASSES = (
     OpenURL,
     ImportWeights,
     RestoreDefaultPresets,
-    Checkpoint,
+    CheckpointGroup,
     LinkCheckpoint,
     UnlinkCheckpoint,
     PREFERENCES_UL_CheckpointList,
