+++ generator_process/__init__.py
@@ -14,7 +14,7 @@ class Generator(Actor):
     from .actions.upscale import upscale
     from .actions.depth_to_image import depth_to_image
     from .actions.control_net import control_net
-    from .actions.huggingface_hub import hf_snapshot_download, hf_list_models, hf_list_installed_models, set_checkpoint_links
+    from .actions.huggingface_hub import hf_snapshot_download, hf_list_models, hf_list_installed_models
     from .actions.ocio_transform import ocio_transform
     from .actions.convert_original_stable_diffusion_to_diffusers import convert_original_stable_diffusion_to_diffusers
     from .actions.detect_seamless import detect_seamless
