+++ generator_process/actions/depth_to_image.py
@@ -5,14 +5,14 @@ from contextlib import nullcontext
 from numpy.typing import NDArray
 import numpy as np
 import random
-from .prompt_to_image import Scheduler, Optimizations, StepPreviewMode, ImageGenerationResult, _configure_model_padding
+from .prompt_to_image import Checkpoint, Scheduler, Optimizations, StepPreviewMode, ImageGenerationResult, _configure_model_padding
 from ...api.models.seamless_axes import SeamlessAxes
 from ..future import Future
 
 def depth_to_image(
     self,
     
-    model: str,
+    model: str | Checkpoint,
 
     scheduler: Scheduler,
