+++ generator_process/models/checkpoint.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+from .model_config import ModelConfig
+
+
+@dataclass
+class Checkpoint:
+    path: str
+    config: ModelConfig | str | None
